# swagger.yaml
openapi: 3.0.0
info:
  title: ramen-api-server
  description: ラーメンマップアプリ用 API サーバー（YouTube / Google Places キャッシュプロキシ）
  version: 1.0.0

servers:
  - url: https://ramen-api-server-dda6frdjhpehcubw.japanwest-01.azurewebsites.net
    description: Azure (Production)
  - url: http://localhost:3000
    description: ローカル開発

tags:
  - name: General
    description: サーバー状態確認
  - name: YouTube
    description: YouTube Data API キャッシュプロキシ
  - name: Places
    description: Google Places API キャッシュプロキシ
  - name: Cache
    description: キャッシュ管理

paths:
  /:
    get:
      tags: [General]
      summary: ヘルスチェック
      responses:
        '200':
          description: サーバー正常動作中
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: ramen-api-server is running
                  cache:
                    type: string
                    example: 3 entries

  /youtube/search:
    get:
      tags: [YouTube]
      summary: YouTube 動画検索（7日キャッシュ）
      description: |
        YouTube Data API の search.list をプロキシする。
        APIキーはサーバー側で管理するためクライアントからの指定不要。
        同一クエリは7日間キャッシュされ API クォータを節約する。
      parameters:
        - name: q
          in: query
          required: true
          description: 検索キーワード（例：SUSURU TV. 一蘭）
          schema:
            type: string
        - name: type
          in: query
          description: 検索タイプ
          schema:
            type: string
            default: video
        - name: maxResults
          in: query
          description: 最大取得件数
          schema:
            type: integer
            default: 5
        - name: order
          in: query
          description: 並び順
          schema:
            type: string
            enum: [relevance, date, viewCount, rating]
            default: relevance
        - name: part
          in: query
          description: レスポンスに含めるパーツ
          schema:
            type: string
            default: snippet
      responses:
        '200':
          description: YouTube search.list レスポンス
          headers:
            X-Cache:
              description: HIT=キャッシュ返却 / MISS=API呼び出し
              schema:
                type: string
                enum: [HIT, MISS]
        '400':
          description: q パラメータが未指定
        '503':
          description: サーバーの YOUTUBE_API_KEY が未設定

  /youtube/videos:
    get:
      tags: [YouTube]
      summary: YouTube 動画詳細取得（30日キャッシュ）
      description: |
        YouTube Data API の videos.list をプロキシする。
        同一 ID リストは30日間キャッシュされる。
      parameters:
        - name: id
          in: query
          required: true
          description: 動画IDのカンマ区切りリスト
          schema:
            type: string
            example: ABC123,DEF456
        - name: part
          in: query
          description: レスポンスに含めるパーツ
          schema:
            type: string
            default: snippet,contentDetails
      responses:
        '200':
          description: YouTube videos.list レスポンス
          headers:
            X-Cache:
              description: HIT=キャッシュ返却 / MISS=API呼び出し
              schema:
                type: string
                enum: [HIT, MISS]
        '400':
          description: id パラメータが未指定
        '503':
          description: サーバーの YOUTUBE_API_KEY が未設定

  /places/nearbysearch:
    get:
      tags: [Places]
      summary: 周辺ラーメン店検索（24時間キャッシュ）
      description: |
        Google Places API の nearbysearch をプロキシする。
        location+radius の組み合わせをタイル単位でキャッシュ（約890m単位）。
        pagetoken 指定時はキャッシュしない（トークンは使い捨て）。
      parameters:
        - name: location
          in: query
          description: 緯度,経度（例：35.7,139.7）
          schema:
            type: string
        - name: radius
          in: query
          description: 検索半径（メートル）
          schema:
            type: integer
            default: 1500
        - name: keyword
          in: query
          schema:
            type: string
            default: ラーメン
        - name: language
          in: query
          schema:
            type: string
            default: ja
        - name: pagetoken
          in: query
          description: 次ページトークン（指定時はキャッシュしない）
          schema:
            type: string
      responses:
        '200':
          description: Places nearbysearch レスポンス
          headers:
            X-Cache:
              schema:
                type: string
                enum: [HIT, MISS, SKIP]

  /places/details:
    get:
      tags: [Places]
      summary: 店舗詳細取得（7日キャッシュ）
      parameters:
        - name: place_id
          in: query
          required: true
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
            default: place_id,name,rating,user_ratings_total,formatted_address,photos,opening_hours,geometry
        - name: language
          in: query
          schema:
            type: string
            default: ja
      responses:
        '200':
          description: Places details レスポンス

  /places/autocomplete:
    get:
      tags: [Places]
      summary: 店舗名オートコンプリート（24時間キャッシュ）
      parameters:
        - name: input
          in: query
          required: true
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            default: ja
      responses:
        '200':
          description: Places autocomplete レスポンス

  /places/photo:
    get:
      tags: [Places]
      summary: 店舗写真（APIキーを隠してリダイレクト）
      parameters:
        - name: photo_reference
          in: query
          required: true
          schema:
            type: string
        - name: maxwidth
          in: query
          schema:
            type: integer
            default: 400
      responses:
        '302':
          description: Google の写真URLへリダイレクト

  /cache/stats:
    get:
      tags: [Cache]
      summary: キャッシュ状況確認
      description: 現在メモリに保持しているキャッシュの一覧と残り有効期限を返す。
      responses:
        '200':
          description: キャッシュ統計
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 3
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        expiresIn:
                          type: string
                          example: 604799s